16-Bit BEEtag
Adaptation of the original 25-bit BEEtag system to reduce the tag complexity to 16 bits. The information stored within each tag is encoded as a 4x4 grid of black and white pixels, surrounded by a white pixel border and a black pixel border. This matrix of pixels encodes 16-bits of information, divided into a 12-bit identification code and a 4-bit error check. The identification code is encoded column-wise within a 4×3 section of the tag pixel matrix as a 12-digit binary number left-padded with zeros. The error check is encoded within the final 4-digit column of the matrix. The first three digits of the error check encode the even parity (1 (white) for an odd number of 1s; 0 (black) for an even number of 1s) of each of the three columns of the identification code. The final digit encodes the parity of the entire identification code. The number of unique identification codes that can be generated from 12 bits is 4096. The number of codes is limited to only those that could operate in a single orientation, leaving 3368 unique codes. The final list of codes used each has a hamming distance of 2 relative to all other codes, which leaves 328 robust codes to be used.

The 'locate16BitCodes.m' function was adapted from an old version of the 'locateCodes.m' function. This means the 'locate16BitCodes.m' function runs much slower than the newer 'locateCodes.m' function.